+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 ## Changelog
-### 0.3.0 (unreleased)
+### 0.3.0
 - Added functionality to detect submitted password forms and save them to the database ([#6](https://github.com/eglia/ff-oc-passwords/issues/6))
 - Added functionality to copy passwords to clipboard ([#7](https://github.com/eglia/ff-oc-passwords/issues/7))
 - Added functionality to fine tune site matching ([#8](https://github.com/eglia/ff-oc-passwords/issues/8))
 - Added proper cleanup function
-- Added warning when using a passwords app older than v17
+- Added warning when using old versions of the password app
 - Improved JSON escape
 - Fixed warnings not showing when opening settings panel
 - Fixed refresh button getting stuck